Move DAGs table to a reusable DataTable component
